+++ src/main/java/cart/application/OrderService.java
@@ -7,9 +7,9 @@ import cart.domain.OrderItem;
 import cart.repository.CartItemRepository;
 import cart.repository.MemberRepository;
 import cart.repository.OrderRepository;
-import cart.ui.dto.OrderCreateRequest;
-import cart.ui.dto.OrderItemDto;
-import cart.ui.dto.OrderShowResponse;
+import cart.ui.dto.request.OrderCreateRequest;
+import cart.ui.dto.response.OrderItemDto;
+import cart.ui.dto.response.OrderShowResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
