+++ src/main/java/cart/ui/CartItemApiController.java
@@ -3,9 +3,9 @@ package cart.ui;
 import cart.application.CartItemService;
 import cart.auth.Auth;
 import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.ui.dto.request.CartItemQuantityUpdateRequest;
+import cart.ui.dto.request.CartItemRequest;
+import cart.ui.dto.response.CartItemResponse;
 import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
