+++ src/main/java/cart/ui/OrderController.java
@@ -3,8 +3,8 @@ package cart.ui;
 import cart.application.OrderService;
 import cart.auth.Auth;
 import cart.domain.Member;
-import cart.ui.dto.OrderCreateRequest;
-import cart.ui.dto.OrderShowResponse;
+import cart.ui.dto.request.OrderCreateRequest;
+import cart.ui.dto.response.OrderShowResponse;
 import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
