+++ src/main/java/cart/ui/ProductApiController.java
@@ -1,8 +1,8 @@
 package cart.ui;
 
 import cart.application.ProductService;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.ui.dto.request.ProductRequest;
+import cart.ui.dto.response.ProductResponse;
 import java.net.URI;
 import java.util.List;
 import org.springframework.http.HttpStatus;
