+++ src/main/java/cart/ui/dto/request/CartItemQuantityUpdateRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.ui.dto.request;
 
 public class CartItemQuantityUpdateRequest {
     private int quantity;
