+++ src/main/java/cart/ui/dto/request/OrderCreateRequest.java
@@ -1,4 +1,4 @@
-package cart.ui.dto;
+package cart.ui.dto.request;
 
 import java.util.List;
