+++ src/main/java/cart/ui/dto/response/CartItemResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.ui.dto.response;
 
 import cart.domain.CartItem;
