+++ src/main/java/cart/ui/dto/response/OrderItemDto.java
@@ -1,20 +1,23 @@
-package cart.dto;
+package cart.ui.dto.response;
 
-import cart.domain.Product;
+import cart.domain.OrderItem;
 
-public class ProductResponse {
+public class OrderItemDto {
+
     private Long id;
     private String name;
     private int price;
     private String imageUrl;
+    private int quantity;
     private int discountRate;
     private int discountedPrice;
 
-    public ProductResponse(
+    private OrderItemDto(
             final Long id,
             final String name,
             final int price,
             final String imageUrl,
+            final int quantity,
             final int discountRate,
             final int discountedPrice
     ) {
@@ -22,18 +25,20 @@ public class ProductResponse {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
+        this.quantity = quantity;
         this.discountRate = discountRate;
         this.discountedPrice = discountedPrice;
     }
 
-    public static ProductResponse of(final Product product) {
-        return new ProductResponse(
-                product.getId(),
-                product.getName(),
-                product.getPrice(),
-                product.getImageUrl(),
-                product.getDiscountRate(),
-                product.calculateDiscountedPrice()
+    public static OrderItemDto from(final OrderItem orderItem) {
+        return new OrderItemDto(
+                orderItem.getId(),
+                orderItem.getName(),
+                orderItem.getPrice(),
+                orderItem.getImageUrl(),
+                orderItem.getQuantity(),
+                orderItem.getDiscountRate(),
+                orderItem.calculateDiscountedPrice()
         );
     }
 
@@ -51,6 +56,10 @@ public class ProductResponse {
 
     public String getImageUrl() {
         return imageUrl;
+    }
+
+    public int getQuantity() {
+        return quantity;
     }
 
     public int getDiscountRate() {
