+++ src/test/java/cart/integration/CartItemIntegrationTest.java
@@ -4,11 +4,11 @@ import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
-import cart.dto.ProductRequest;
 import cart.repository.MemberRepository;
+import cart.ui.dto.request.CartItemQuantityUpdateRequest;
+import cart.ui.dto.request.CartItemRequest;
+import cart.ui.dto.request.ProductRequest;
+import cart.ui.dto.response.CartItemResponse;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.Arrays;
