+++ src/test/java/cart/integration/ProductIntegrationTest.java
@@ -3,8 +3,8 @@ package cart.integration;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.ui.dto.request.ProductRequest;
+import cart.ui.dto.response.ProductResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
